fix: Correcion del navbar, formularios y creacion de notificaciones
