+++ frent/src/App.css
@@ -25,14 +25,10 @@
 
 body{
   font-family: "Alata", sans-serif;
-  
+  background-color: #fcfcfc;  
 }
 
-.body-app{
-  display: flex;
-  flex-flow: row;
-  justify-content: center;
-  background-color: #fcfcfc;
-  padding: 30px 0px;
-}
+/* .body-app{
+  background-color: #fcfcfc;  
+} */
