+++ frent/src/App.jsx
@@ -2,13 +2,15 @@ import {BrowserRouter, Routes, Route} from 'react-router-dom'
 import './App.css';
 /* import { CustomerFormPage } from './components/form/customerFormPage'; */
 /* import { CustomerFormFriend } from './components/form/customeFormFriend'; */
-import NavBarN from './components/Navbar/NavBarN';
+import NavBar from './components/Navbar/NavBar';
 import Home from './pages/home/Home';
-import ListFriend from './pages/listFriend/VeryFriends';
+import ListFriend from './pages/listFriend/ListFriends';
 import SelectionRegister from './pages/registros/SelectionRegister';
 import FriendProfile from './pages/listFriend/FriendProfile';
-import { Test } from './components/Forms/CustomerForm/test';
-import RentaForm from './components/Forms/CustomerForm/rentFriends/RentaForm';
+import { CustomerForm } from './components/Forms/CustomerForm/CustomerForm';
+import RentFriendForm from './components/Forms/rentFriends/RentaForm';
+import { FriendForm } from './components/Forms/FriendForm/FriendForm';
+import RentalSection from './pages/rentalSection/RentalSection';
 
 
 
@@ -18,15 +20,16 @@ function App() {
     <>
     <div className='body-app'>
     <BrowserRouter>
-      <NavBarN></NavBarN>
+      <NavBar></NavBar>
       <Routes>
-        <Route path="/home" element={<Home/>}/>
-        <Route path="/form" element={<SelectionRegister></SelectionRegister>}/>
-        <Route path="/cliente" element={<Test></Test>}/>
-        {/* <Route path="/friend" element={<CustomerFormFriend></CustomerFormFriend>}/> */}
-        <Route path='/listFriend' element = {<ListFriend></ListFriend>}/>
-        <Route path="listFriend/:id" element={<FriendProfile></FriendProfile>} />
-        <Route path='/rentaForm' element = {<RentaForm></RentaForm>}></Route>
+        <Route path="/" element={<Home/>}/>
+        <Route path="/form" element={<SelectionRegister/>}/>
+        <Route path="/customer" element={<CustomerForm/>}/>
+        <Route path="/friend" element={<FriendForm/>} />
+        <Route path="/listFriend" element = {<ListFriend/>}/>
+        <Route path="listFriend/:id" element={<FriendProfile/>} />
+        <Route path="/rentaForm" element = {<RentFriendForm/>}/>
+        <Route path="/rentalSectio" element ={<RentalSection/>} />
         <Route path="/*" element={<h1 className='text-center'>404 Page Not Found</h1>}></Route>
       </Routes>
     </BrowserRouter>
