+++ frent/src/components/Forms/CustomerForm/customerForm.css
@@ -1,86 +1,91 @@
-.FormularioRegistro{
-  display: flex;
-  justify-content: center;
-  padding: 30px 0px;
-}
-
-.container-principal{
-  background: #fff;
-  width: 55%;
-  border-radius: 12px;
-  font-weight: 400; /*quitar*/
-  font-style: normal;
-  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-input {
-  border: 2px solid #ccc;
-  border-radius: 4px;
-}
-
-input::placeholder{
-  color: #7b7b7b;
-}
-
-select {
-  border: 2px solid #ccc;
-  border-radius: 4px; 
-}
-
-#formulario-cliente {
+.body-page{
     display: flex;
+    flex-flow: row;
     justify-content: center;
-  }
-  
-  #formulario-cliente label {
-    margin-bottom: 5px;
-  }
-  
-  .container-principal > h2 {
+    padding: 24px 0px;
+}
+
+.form-body-container{
+    padding: 24px 48px;
+    width: 55%;
+    border-radius: 12px;
+    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
+    background-color: #fff;
+}
+
+.descripction{
+    display: flex;
+    flex-flow: column wrap;
+    gap: 8px;
+}
+
+.descripction > textarea{
+    border: 3px solid #ccc;
+    border-radius: 8px;
+}
+
+.location {
+    display: flex;
+    flex-flow: column wrap;
+    gap: 8px;
+}
+
+.location > textarea{
+    border: 3px solid #ccc;
+    border-radius: 8px;
+}
+
+.error-message{
+    font-size: var(--pM-size);
+    color: red;
+    margin-bottom: 8px;
+}
+
+
+.form-body-container > h3 {
+    font-size: var(--h3-size);
     text-align: center;
-  }
-  
-  .container {
+}
+
+.colums-inputs {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr)); /* Ajusta el ancho de las columnas según el contenido */
+    gap: 16px; /* Espacio entre elementos */
+}
+
+.input-1c,
+.input-2c,
+.input-4c {
+    width: 100%;
+}
+
+.pais-select{
+    width: 100%;
+}
+.input-1c {
+    grid-column: span 1; /* Ocupa una columna completa */
+}
+
+.input-2c {
+    grid-column: span 2; /* Ocupa dos columnas */
+}
+
+.input-4c {
+    grid-column: span 4; /* Ocupa cuatro columnas */
+}
+
+.textAreaDescription {
+    width: 100%;
+    height: 100px; /* Ajusta la altura según sea necesario */
+}
+
+.textArealocation {
+    width: 100%;
+    height: 100px;
+}
+
+.buttons-section {
     display: flex;
-    justify-content: center;
-    flex-wrap: wrap;
-  }
-  
-  .campos-form {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .campos-mitad {
-    width: 42%;
-    margin: 10px 10px;
-  }
-  
-  .campos-mitad-mitad {
-    width: 20%;
-    margin: 10px 10px;
-  }
-  
-  .campo-full {
-    width: 86%;
-  }
-  
-  input,
-  select {
-    border-radius: 5px;
-    padding: 10px;
-  }
-  
-  .llevar-inicio {
-    margin-right: 67%;
-  }
-
-  
-  
-  span{
-    color: tomato;
-    font-size: x-small;
-  }
-
-
-  
+    justify-content: space-between;
+    margin-top: 48px;
+}
