+++ frent/src/components/Forms/rentFriends/RentaForm.css
