+++ frent/src/components/Navbar/Navbar.css
@@ -1,0 +1,47 @@
+.logo-img{
+  width: 120px;
+}
+
+.navbar-body{
+  display: flex;
+  flex-flow: row wrap;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0px 48px;
+  
+  background-color: var(--primary-color);
+  width: 100% vh;
+}
+
+.navbar-options-list{
+  display: flex;
+  flex-flow: row wrap;
+  align-items: center;
+}
+
+.navbar-option{
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 8px 12px;
+  border-radius: 4px;
+  border: none;
+  background-color: var(--primary-color);
+  color: #000;
+}
+
+.navbar-option:hover{
+  background-color: var(--secondary-color);
+  color: #fff !important;
+}
+
+/*
+.a-option:hover{
+  color: #fff;
+} */
+
+a{
+  text-decoration: none;
+  font-size: var(--pL-size);
+  font-weight: 500;
+}
