--- frent/src/components/Navbar/NavbarU.css
@@ -1,30 +1,0 @@
-nav {
-    background-color: #333; 
-    padding: 10px 20px; 
-  }
-  
-  ul {
-    list-style-type: none;
-    padding: 0;
-    margin: 0;
-  }
-  
-  li {
-    display: inline; 
-    margin-right: 20px; 
-  }
-  
-  li:last-child {
-    margin-right: 0;
-  }
-  
-  a {
-    text-decoration: none; 
-    color: #fff; 
-    font-weight: bold; 
-  }
-  
-  a:hover {
-    color: #ffcc00; 
-  }
-
