+++ frent/src/pages/home/Home.jsx
@@ -1,7 +1,14 @@
 import React from 'react'
+const centrar = {
+  textAlign: 'center'
+};
 
 export default function Home() {
   return (
-    <div>Home</div>
+    <div style={centrar}>
+      <h1>
+        Home
+      </h1> 
+    </div>
   )
 }
