+++ frent/src/pages/listFriend/ListFriend.css
@@ -11,10 +11,10 @@
     max-width: 1100px;
     margin: auto;
 }
-.lista:hover {
+/* .lista:hover {
     box-shadow: 0 8px 16px rgba(0, 0, 255, 0.4); 
     transition: cubic-bezier(0.165, 0.84, 0.44, 1) ease-in-out;
-  }
+  } */
   
   .card {
     border: 1px solid #ccc;
@@ -53,16 +53,16 @@
 .btnC {
     display: inline-block;
     padding: 10px 20px;
-    background-color: #007bff;
+    background-color: var(--primary-color);
     color: #fff;
     border-radius: 5px;
     text-decoration: none;
     transition: background-color 0.3s ease;
   }
   
-  .btnC:hover {
+  /* .btnC:hover {
     background-color: #0056b3;
-  }
+  } */
