+++ frent/src/pages/registros/SelectionRegister.css
@@ -1,34 +1,46 @@
 .selection-register-container{
     display: grid;
     place-items: center;
-    margin: 40px;
+    margin: 0px 40px;
+    font-family: Alata;
 }
 .selection-register-container h1 {
-    font-size: 52px;
+    font-size: var(--h1-size);
     text-align: start;
-    font-family: Arial, Helvetica, sans-serif;
 }
 
 .selection-register-container p {
-    margin-top: 40px;
+    margin-top: 16px;
     color: #737373;
     font-size: larger;
-    display: flex;
-    justify-content: center;
     text-align: center;
 }
 
-.botonRegister {
-    display: inline-block;
-    margin: 10px;
-    padding: 10px 20px;
-    text-decoration: none;
-    color: white;
-    background-color: #007bff;
-    border-radius: 5px;
-  }
-  
-  .botonRegister:hover {
-    background-color: #0056b3;
-  }
-  
+.register-buttons {
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: space-between;
+    gap: 48px;
+}
+
+
+.register-btn {
+    background-color: var(--tertiary-color);
+    border-radius: 8px;
+}
+
+.button-paragraph{
+    color: #fff!important;
+}
+
+.register-btn{
+    margin-top: 16px;
+    display: flex;
+    flex-flow: column wrap;
+    align-items: center;
+    padding: 24px;
+}
+.register-btn > img {
+    width: 160px;
+    height: auto;
+}
