+++ frent/src/pages/rentalSection/RentalSection.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function RentalSection() {
+  return (
+    <div>RentalSection</div>
+  );
+}
